Rename ScriptSetup to ScriptTemplateSubpage
