+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/PageTemplateView/index.tsx
