+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/PageTemplateView/useStyles.ts
