+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/ScriptTemplateGradebox/index.tsx
