+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/ScriptTemplateGradebox/useStyles.ts
