+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/ScriptTemplateView/QuestionTemplateSelect.tsx
@@ -2,6 +2,8 @@ import React from "react";
 import { TreeSelect } from "antd";
 import { QuestionTemplateTreeData } from "backend/src/types/questionTemplates";
 import useScriptSetup from "AuthenticatedApp/paperSetup/context/ScriptSetupContext";
+import { useFormikContext } from "formik";
+import { NewQuestionTemplateValues } from "./QuestionTemplateDialog";
 
 interface TreeData {
   value: string;
@@ -11,8 +13,6 @@ interface TreeData {
 }
 
 interface SelectTreeProps {
-  initialValue: string;
-  onChange: (value: any) => void;
   container: (triggerNode: any) => any;
 }
 
@@ -30,7 +30,9 @@ const GenerateTreeData = (
 };
 
 const QuestionTemplateSelect: React.FC<SelectTreeProps> = props => {
-  const [value, setValue] = React.useState(props.initialValue);
+  const { values, setFieldValue } = useFormikContext<
+    NewQuestionTemplateValues
+  >();
   const { scriptTemplateSetupData } = useScriptSetup();
   const treeData = scriptTemplateSetupData.questionTemplates.map(d =>
     GenerateTreeData(d)
@@ -39,14 +41,13 @@ const QuestionTemplateSelect: React.FC<SelectTreeProps> = props => {
     <TreeSelect
       showSearch
       style={{ width: "100%" }}
-      value={value}
+      value={values.parentName}
       dropdownStyle={{ maxHeight: 400, overflow: "auto" }}
       placeholder="Please select"
       allowClear
       treeDefaultExpandAll
       onChange={(v: string) => {
-        setValue(v);
-        props.onChange(v);
+        setFieldValue("parentName", v);
       }}
       treeData={treeData}
       getPopupContainer={props.container}
