+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/ScriptTemplateView/index.tsx
