+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/ScriptTemplateView/useStyles.ts
