+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/index.tsx
