+++ frontend/src/AuthenticatedApp/paperSetup/ScriptTemplateSubpage/styles.css
